+++ src/ui/components/AnimatedIcon.native.test.tsx
@@ -1,12 +1,12 @@
 import React from 'react'
-import { Animated } from 'react-native'
+// eslint-disable-next-line no-restricted-imports
+import { Animated, TouchableOpacity } from 'react-native'
 
-import { fireEvent, render, screen, waitFor } from 'tests/utils/web'
+import { fireEvent, render, screen, waitFor } from 'tests/utils'
 import { theme } from 'theme'
-import { TouchableOpacity } from 'ui/components/TouchableOpacity'
+import { Logo } from 'ui/svg/icons/Logo'
 
-import { Logo } from '../../svg/icons/Logo'
-import { AnimatedIcon } from '../AnimatedIcon'
+import { AnimatedIcon } from './AnimatedIcon'
 
 const DummyComponent: React.FC = () => {
   const animatedValue = new Animated.Value(0)
@@ -30,22 +30,25 @@ const DummyComponent: React.FC = () => {
 describe('AnimatedIcon', () => {
   it('should display only the first color before animation', () => {
     render(<DummyComponent />)
+
     const initialContainer = screen.getByTestId('initial-icon-container')
     const finalContainer = screen.getByTestId('final-icon-container')
 
-    expect(initialContainer.style.opacity).toBe('1')
-    expect(finalContainer.style.opacity).toBe('0')
+    expect(initialContainer.props.style.opacity).toBe(1)
+    expect(finalContainer.props.style.opacity).toBe(0)
   })
 
-  it('should display only the last color after animation [Web Integration]', async () => {
+  it('should display only the last color after animation', async () => {
     render(<DummyComponent />)
-    fireEvent.click(screen.getByTestId('dummyPressable'))
+
+    fireEvent.press(screen.getByTestId('dummyPressable'))
+
     const initialContainer = screen.getByTestId('initial-icon-container')
     const finalContainer = screen.getByTestId('final-icon-container')
-    await waitFor(() => {
-      expect(finalContainer.style.opacity).toBe('1')
-    })
 
-    expect(initialContainer.style.opacity).toBe('0')
+    await waitFor(() => {
+      expect(finalContainer.props.style.opacity).toBe(1)
+      expect(initialContainer.props.style.opacity).toBe(0)
+    })
   })
 })
