+++ src/ui/components/AnimatedIcon.web.test.tsx
@@ -1,12 +1,12 @@
 import React from 'react'
-// eslint-disable-next-line no-restricted-imports
-import { Animated, TouchableOpacity } from 'react-native'
+import { Animated } from 'react-native'
 
-import { fireEvent, render, screen, waitFor } from 'tests/utils'
+import { fireEvent, render, screen, waitFor } from 'tests/utils/web'
 import { theme } from 'theme'
-import { Logo } from 'ui/svg/icons/Logo'
+import { TouchableOpacity } from 'ui/components/TouchableOpacity'
 
-import { AnimatedIcon } from '../AnimatedIcon'
+import { Logo } from '../svg/icons/Logo'
+import { AnimatedIcon } from './AnimatedIcon'
 
 const DummyComponent: React.FC = () => {
   const animatedValue = new Animated.Value(0)
@@ -30,25 +30,22 @@ const DummyComponent: React.FC = () => {
 describe('AnimatedIcon', () => {
   it('should display only the first color before animation', () => {
     render(<DummyComponent />)
-
     const initialContainer = screen.getByTestId('initial-icon-container')
     const finalContainer = screen.getByTestId('final-icon-container')
 
-    expect(initialContainer.props.style.opacity).toBe(1)
-    expect(finalContainer.props.style.opacity).toBe(0)
+    expect(initialContainer.style.opacity).toBe('1')
+    expect(finalContainer.style.opacity).toBe('0')
   })
 
-  it('should display only the last color after animation', async () => {
+  it('should display only the last color after animation [Web Integration]', async () => {
     render(<DummyComponent />)
-
-    fireEvent.press(screen.getByTestId('dummyPressable'))
-
+    fireEvent.click(screen.getByTestId('dummyPressable'))
     const initialContainer = screen.getByTestId('initial-icon-container')
     const finalContainer = screen.getByTestId('final-icon-container')
-
     await waitFor(() => {
-      expect(finalContainer.props.style.opacity).toBe(1)
-      expect(initialContainer.props.style.opacity).toBe(0)
+      expect(finalContainer.style.opacity).toBe('1')
     })
+
+    expect(initialContainer.style.opacity).toBe('0')
   })
 })
