+++ src/ui/components/ImageTile.native.test.tsx
